+++ common/buildcraft/core/triggers/CraftingHelper.java
